chore: move interfaces to reducer and import on screen
