+++ app/reducers/selectedUsers.ts
@@ -1,8 +1,17 @@
-import { IUser, ApplicationActions } from '../definitions';
+import { ApplicationActions } from '../definitions';
 import { SELECTED_USERS } from '../actions/actionsTypes';
 
+export interface ISelectedUser {
+	_id: string;
+	name: string;
+	fname: string;
+	search?: boolean;
+	// username is used when is from searching
+	username?: string;
+}
+
 export interface ISelectedUsers {
-	users: IUser[];
+	users: ISelectedUser[];
 	loading: boolean;
 }
